// packages/wings/src/proto/wings/flight/ingestion.ts
export enum IngestionOperation {
  UNSPECIFIED = 0,
  UPSERT = 1,
  DELETE = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface IngestionRequestMetadata {
  $type: "wings.flight.IngestionRequestMetadata";
  readonly requestId: bigint;
  readonly operation: IngestionOperation;
  readonly partitionValue: PartitionValue | undefined;
}

export interface IngestionResponseMetadata {
  $type: "wings.flight.IngestionResponseMetadata";
  readonly requestId: bigint;
  readonly accepted: boolean;
  readonly message: string;
}

function createBaseIngestionRequestMetadata(): IngestionRequestMetadata {
  return {
    $type: "wings.flight.IngestionRequestMetadata",
    requestId: 0n,
    operation: 0,
    partitionValue: undefined,
  };
}

export const IngestionRequestMetadata: MessageFns<
  IngestionRequestMetadata,
  "wings.flight.IngestionRequestMetadata"
> = {
  $type: "wings.flight.IngestionRequestMetadata" as const,

  encode(
    message: IngestionRequestMetadata,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.requestId !== 0n) {
      if (BigInt.asUintN(64, message.requestId) !== message.requestId) {
        throw new globalThis.Error(
          "value provided for field message.requestId of type uint64 too large",
        );
      }
      writer.uint32(8).uint64(message.requestId);
    }
    if (message.operation !== 0) {
      writer.uint32(16).int32(message.operation);
    }
    if (message.partitionValue !== undefined) {
      PartitionValue.encode(message.partitionValue, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IngestionRequestMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIngestionRequestMetadata() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.requestId = reader.uint64() as bigint;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.operation = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.partitionValue = PartitionValue.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IngestionRequestMetadata {
    return {
      $type: IngestionRequestMetadata.$type,
      requestId: isSet(object.requestId) ? BigInt(object.requestId) : 0n,
      operation: isSet(object.operation) ? ingestionOperationFromJSON(object.operation) : 0,
      partitionValue: isSet(object.partitionValue)
        ? PartitionValue.fromJSON(object.partitionValue)
        : undefined,
    };
  },

  toJSON(message: IngestionRequestMetadata): unknown {
    const obj: any = {};
    if (message.requestId !== 0n) {
      obj.requestId = message.requestId.toString();
    }
    if (message.operation !== 0) {
      obj.operation = ingestionOperationToJSON(message.operation);
    }
    if (message.partitionValue !== undefined) {
      obj.partitionValue = PartitionValue.toJSON(message.partitionValue);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<IngestionRequestMetadata>, I>>(
    base?: I,
  ): IngestionRequestMetadata {
    return IngestionRequestMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IngestionRequestMetadata>, I>>(
    object: I,
  ): IngestionRequestMetadata {
    const message = createBaseIngestionRequestMetadata() as any;
    message.requestId = object.requestId ?? 0n;
    message.operation = object.operation ?? 0;
    message.partitionValue =
      object.partitionValue !== undefined && object.partitionValue !== null
        ? PartitionValue.fromPartial(object.partitionValue)
        : undefined;
    return message;
  },
};

messageTypeRegistry.set(IngestionRequestMetadata.$type, IngestionRequestMetadata);

function createBaseIngestionResponseMetadata(): IngestionResponseMetadata {
  return {
    $type: "wings.flight.IngestionResponseMetadata",
    requestId: 0n,
    accepted: false,
    message: "",
  };
}

export const IngestionResponseMetadata: MessageFns<
  IngestionResponseMetadata,
  "wings.flight.IngestionResponseMetadata"
> = {
  $type: "wings.flight.IngestionResponseMetadata" as const,

  encode(
    message: IngestionResponseMetadata,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.requestId !== 0n) {
      if (BigInt.asUintN(64, message.requestId) !== message.requestId) {
        throw new globalThis.Error(
          "value provided for field message.requestId of type uint64 too large",
        );
      }
      writer.uint32(8).uint64(message.requestId);
    }
    if (message.accepted !== false) {
      writer.uint32(16).bool(message.accepted);
    }
    if (message.message !== "") {
      writer.uint32(26).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IngestionResponseMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIngestionResponseMetadata() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.requestId = reader.uint64() as bigint;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.accepted = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IngestionResponseMetadata {
    return {
      $type: IngestionResponseMetadata.$type,
      requestId: isSet(object.requestId) ? BigInt(object.requestId) : 0n,
      accepted: isSet(object.accepted) ? globalThis.Boolean(object.accepted) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
    };
  },

  toJSON(message: IngestionResponseMetadata): unknown {
    const obj: any = {};
    if (message.requestId !== 0n) {
      obj.requestId = message.requestId.toString();
    }
    if (message.accepted !== false) {
      obj.accepted = message.accepted;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<IngestionResponseMetadata>, I>>(
    base?: I,
  ): IngestionResponseMetadata {
    return IngestionResponseMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IngestionResponseMetadata>, I>>(
    object: I,
  ): IngestionResponseMetadata {
    const message = createBaseIngestionResponseMetadata() as any;
    message.requestId = object.requestId ?? 0n;
    message.accepted = object.accepted ?? false;
    message.message = object.message ?? "";
    return message;
  },
};

messageTypeRegistry.set(IngestionResponseMetadata.$type, IngestionResponseMetadata);

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends { readonly $case: string }
        ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { readonly $case: T["$case"] }
        : T extends {}
          ? { [K in Exclude<keyof T, "$type">]?: DeepPartial<T[K]> }
          : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P> | "$type">]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T, V extends string> {
  readonly $type: V;
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}